+++ exporter/utils.py
@@ -2,15 +2,6 @@ import inspect
 import logging
 from datetime import datetime
 
-from hca_ingest.utils.date import parse_date_string
-
-DCP_VERSION_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
-
-
-def to_dcp_version(date_str: str):
-    date = parse_date_string(date_str)
-    return date.strftime(DCP_VERSION_FORMAT)
-
 
 def exec_time(logger: logging.Logger, level: int):
     def _exec_time(func):
